requirements.txt
tensorflow
streamlit
opencv-python-headless
numpy
pandas
pillow
requests
scikit-learn
